refactor: some file restructure feat: add check into process_shcedule_commit
