+++ programs/magicblock/src/schedule_transactions/mod.rs
@@ -11,7 +11,7 @@ pub(crate) mod transaction_scheduler;
 
 use std::sync::Arc;
 
-use magicblock_core::intent::CommittedAccount;
+use magicblock_core::intent::types::CommittedAccount;
 use magicblock_magic_program_api::{
     pda::CALLBACK_SIGNER, MAGIC_CONTEXT_PUBKEY,
 };
