+++ programs/magicblock/src/schedule_transactions/process_schedule_commit_tests.rs
@@ -1,6 +1,7 @@
 use std::collections::HashMap;
 
 use assert_matches::assert_matches;
+use magicblock_core::intent::{ACTUAL_COMMIT_LIMIT, COMMIT_FEE_LAMPORTS};
 use magicblock_magic_program_api::{
     args::{
         ActionArgs, BaseActionArgs, MagicIntentBundleArgs, ShortAccountMeta,
@@ -21,9 +22,7 @@ use solana_signer::Signer;
 
 use crate::{
     magic_context::MagicContext,
-    magic_scheduled_base_intent::{
-        ScheduledIntentBundle, ACTUAL_COMMIT_LIMIT, COMMIT_FEE_LAMPORTS,
-    },
+    magic_scheduled_base_intent::ScheduledIntentBundle,
     magic_sys::COMMIT_LIMIT,
     schedule_transactions::{
         magic_fee_vault_pubkey, transaction_scheduler::TransactionScheduler,
