+++ test-integration/schedulecommit/test-scenarios/tests/02_commit_and_undelegate.rs
@@ -379,6 +379,12 @@ fn run_test_for_commit_huge_order_book_account(
 
         assert_one_committee_was_committed(&ctx, &res, true);
         match commit_type {
+            ScheduleCommitType::Commit => {
+                assert_one_committee_account_was_not_undelegated_on_chain(&ctx);
+            }
+            ScheduleCommitType::CommitAndUndelegate => {
+                assert_one_committee_account_was_undelegated_on_chain(&ctx);
+            }
             ScheduleCommitType::CommitFinalize => {
                 assert_one_committee_account_was_not_undelegated_on_chain(&ctx);
             }
